Add AlterDatabaseDiscoveryRuleStatementUpdater
